TEST-31 Added LocalUser and LocalGroup annotations to dominion. (#32)
* TEST-31 Added LocalUser and LocalGroup annotations to dominion.
Modified installation of dominion to create a new dominion local user to perform SSH operations.
Updated internal logic to ensure the dominion features would be added/processed as the last feature when PaxExam starts the container. This ensures that the Dominion PaxExam doesn't come up before any other features.

Modified to use all extension points provided by all registered extension factories. Modified to filter distributions if dominion.distribution is specified allowing for a test to be executed with a different distribution.

Updated to allow interpolation of annotation attributes for options annotated on extensions