TEST-33 Enhancements and fixes.
Fixed Dominion to support finding out the project base dir at runtime in order to load resources from there.
Ensured that Dominion would initialize Aether's package for support of the mvn URL protocol whenever the MavenUrlReference was used to get a URL.
Added support for a new source type allowing a Maven URL to be specified as the source.
Upgraded Aether support to the latest version.Updated Dominion feature to add dependency to Pax aether and google guava bundles.
Added info logging of log levels being set.
Enhanced Dominion to automatically grant itself permissions when a security manager was in play; require registered extension to be supplied by the distro.